usbdriver.h: Add prototype for Is_usbworking
Without this, dpcmd produces "Error: Programmers are not connected."
when trying to do anything other than detection. The call in dpcmd.c
treats the function as returning an int rather than a bool, which causes
checks like Is_usbworking(0)==true to fail. With this change, that check
succeeds, allowing dpcmd to proceed.