Add Meson build backend
This commit adds a meson build backend while retaining the existing Make
backend. It also runs the `make release` command, which reformatted many
of the files, as well as removing a good number of executable permissions.

Building with `meson` can be done by:

```bash
$ meson setup builddir
$ meson compile -C builddir
```

The `dpcmd` executable is now inside `builddir` as `dpcmd`.

Finally, this also adds a `.clang-format` file. This makes it easier for
IDEs to understand the formatting considerations for this repository. Its
contents are simply a `--dump-config` from the standard `WebKit` `clang-format`
configuration.

Instructions on how to build have also been added to the `README.md` file,
which still includes the same instructions for using `make`.