Makefile: add conditional stripping
This patch introduces a `NOSTRIP` conditional flag to the `install`
target in the Makefile. The purpose of this flag is to allow the
disabling of binary stripping during installation. This is
particularly useful in development environments, such as BitBake,
which handle stripping internally.

When `NOSTRIP` is set to `1`, the `strip` command will be skipped
during the installation process, preventing potential conflicts or
redundant operations.

Signed-off-by: Tymoteusz Burak <tymoteusz.burak@3mdeb.com>